perubahan total

+++ app-core/database/migrations/2024_10_14_062159_create_t_p_reviews_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('t_users')->onDelete('cascade');
             $table->foreignId('order_id')->constrained('t_orders')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('t_product')->onDelete('cascade');
-            $table->tinyInteger('rating'); // Rating between 1 and 5
-            $table->text('comment');        
+            $table->integer('rating');
+            $table->text('comment');
             $table->timestamps();
         });
     }
